Move getSiteBlocksCurrentPage to state/reader
