+++ client/state/reader/site-blocks/selectors/get-site-blocks-current-page.js
