+++ client/state/selectors/test/get-site-blocks-current-page.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import getSiteBlocksCurrentPage from 'state/selectors/get-site-blocks-current-page';
+import getSiteBlocksCurrentPage from 'state/reader/site-blocks/selectors/get-site-blocks-current-page';
 
 describe( 'getSiteBlocksCurrentPage()', () => {
 	test( 'should return the current page', () => {
